Clean up, rename LLVMTarget to avoid confusion with core LLVMTarget
