+++ mlir/lib/CAPI/CompilerDriver.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "Catalyst/Driver/LLVMTarget.h"
+#include "Catalyst/Driver/CatalystLLVMTarget.h"
 #include "Catalyst/Driver/Pipelines.h"
 
 #include "Catalyst/IR/CatalystDialect.h"
