+++ mlir/lib/Catalyst/Driver/CatalystLLVMTarget.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "Catalyst/Driver/LLVMTarget.h"
+#include "Catalyst/Driver/CatalystLLVMTarget.h"
 
 #include "mlir/IR/FunctionInterfaces.h"
 #include "mlir/Target/LLVMIR/Dialect/Builtin/BuiltinToLLVMIRTranslation.h"
@@ -24,7 +24,11 @@
 #include "Gradient/IR/GradientDialect.h"
 
 using namespace mlir;
+
 namespace {
+/// Emit the LLVM IR metadata required to register custom gradients in Enzyme.
+/// This interface will convert `gradient.augment` and `gradient.vjp` attributes on function-like
+/// ops to the metadata read by Enzyme.
 class GradientToEnzymeMetadataTranslation : public LLVMTranslationDialectInterface {
     using LLVMTranslationDialectInterface::LLVMTranslationDialectInterface;
 
@@ -59,7 +63,6 @@ class GradientToEnzymeMetadataTranslation : public LLVMTranslationDialectInterfa
                               *llvm::MDNode::get(ctx, llvm::ConstantAsMetadata::get(augmented)));
         function->addMetadata("enzyme_gradient",
                               *llvm::MDNode::get(ctx, llvm::ConstantAsMetadata::get(vjp)));
-        function->dump();
         return success();
     }
 };
